requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.5.0
torchmetrics>=0.6.0

# --------- hydra --------- #
hydra-core==1.1.0
hydra-colorlog==1.1.0
hydra-optuna-sweeper==1.1.0
# hydra-ax-sweeper
# hydra-ray-launcher
# hydra-submitit-launcher

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# torch_tb_profiler

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis

# --------- others --------- #
python-dotenv   # loading env variables from .env file
pytest          # tests
sh              # for running bash commands in some tests
rich            # beautiful text formatting in terminal
termcolor==1.1.0
easy-to-hard-data==0.1.7
icecream==2.1.0
pandas==1.3.5
hydra-core==1.1.1
